searchscree,search,sidebarsearch working
